+++ test/congruences_test.h
@@ -145,7 +145,7 @@
 #define POL_8_SOL_2 1431655765
 
 
-int POL_1_COEFFS[POL_1_DEGREE+1] = {POL_1_COEFF_0,
+long POL_1_COEFFS[POL_1_DEGREE+1] = {POL_1_COEFF_0,
 									POL_1_COEFF_1,
 									POL_1_COEFF_2,
 									POL_1_COEFF_3,
@@ -163,7 +163,7 @@ int POL_1_SOLS[NUM_OF_POL_1_SOLS] = {POL_1_SOL_0,
 									 POL_1_SOL_10,
 									 POL_1_SOL_11};
 
-int POL_2_COEFFS[POL_2_DEGREE+1] = {POL_2_COEFF_0,
+long POL_2_COEFFS[POL_2_DEGREE+1] = {POL_2_COEFF_0,
 									POL_2_COEFF_1,
 									POL_2_COEFF_2,
 									POL_2_COEFF_3,
@@ -196,7 +196,7 @@ int POL_2_SOLS[NUM_OF_POL_2_SOLS] = {POL_2_SOL_0,
 									 POL_2_SOL_22,
 									 POL_2_SOL_23};
 
-int POL_3_COEFFS[POL_3_DEGREE+1] = {POL_3_COEFF_0,
+long POL_3_COEFFS[POL_3_DEGREE+1] = {POL_3_COEFF_0,
 									POL_3_COEFF_1,
 									POL_3_COEFF_2,
 									POL_3_COEFF_3,
@@ -210,7 +210,7 @@ int POL_3_COEFFS[POL_3_DEGREE+1] = {POL_3_COEFF_0,
 
 int * POL_3_SOLS = NULL;
 
-int POL_4_COEFFS[POL_4_DEGREE+1] = {POL_4_COEFF_0,
+long POL_4_COEFFS[POL_4_DEGREE+1] = {POL_4_COEFF_0,
 									POL_4_COEFF_1,
 									POL_4_COEFF_2};
 
@@ -219,7 +219,7 @@ int POL_4_SOLS[NUM_OF_POL_4_SOLS] = {POL_4_SOL_0,
 									POL_4_SOL_2,
 									POL_4_SOL_3};
 
-int POL_5_COEFFS[POL_5_DEGREE+1] = {POL_5_COEFF_0,
+long POL_5_COEFFS[POL_5_DEGREE+1] = {POL_5_COEFF_0,
 									POL_5_COEFF_1,
 									POL_5_COEFF_2,
 									POL_5_COEFF_3,
@@ -242,7 +242,7 @@ int POL_5_SOLS[NUM_OF_POL_5_SOLS] = {POL_5_SOL_0,
 									POL_5_SOL_7,
 									POL_5_SOL_8};
 
-int POL_6_COEFFS[POL_6_DEGREE+1] = {POL_6_COEFF_0,
+long POL_6_COEFFS[POL_6_DEGREE+1] = {POL_6_COEFF_0,
 									POL_6_COEFF_1,
 									POL_6_COEFF_2,
 									POL_6_COEFF_3,
@@ -259,13 +259,13 @@ int POL_6_SOLS[NUM_OF_POL_6_SOLS] = {POL_6_SOL_0,
 									 POL_6_SOL_6,
 									 POL_6_SOL_7};
 
-int POL_7_COEFFS[POL_7_DEGREE+1] = {POL_7_COEFF_0,
+long POL_7_COEFFS[POL_7_DEGREE+1] = {POL_7_COEFF_0,
 										POL_7_COEFF_1,
 									  POL_7_COEFF_2};
 
 int POL_7_SOLS[NUM_OF_POL_7_SOLS] = {POL_7_SOL_0};
 
-int POL_8_COEFFS[POL_8_DEGREE+1] = {POL_8_COEFF_0,
+long POL_8_COEFFS[POL_8_DEGREE+1] = {POL_8_COEFF_0,
 										POL_8_COEFF_1,
 									  POL_8_COEFF_2};
